+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - TextMeshPro support
 - LoadSceneAfterTime
 - Attributes 2.1.0 package dependency
+- Lock/Unlock Loading Scene
 
 ### Changed
 - Scriptable Settings Provider for Scene Manager